refactor(KakaoApi): 여러 도메인에서 사용하므로 패키지 구조 일부 변경
